✨ add config option to create schemafull table/event
