+++ .surrealdb
@@ -1,5 +1,6 @@
 [core]
     path = "./tests-files"
+    schema = "less"
 
 [db]
     url = "localhost:8000"
