+++ src/config/db.rs
@@ -1,5 +1,4 @@
-use ini::Ini;
-use std::path::Path;
+use super::common::retrieve_config_value;
 
 pub struct DbConfig {
     pub url: Option<String>,
@@ -7,23 +6,6 @@ pub struct DbConfig {
     pub password: Option<String>,
     pub ns: Option<String>,
     pub db: Option<String>,
-}
-
-fn load_config() -> Option<Ini> {
-    let surrealdb_config_file = Path::new(".surrealdb");
-    Ini::load_from_file(surrealdb_config_file).ok()
-}
-
-fn retrieve_config_value(section: &str, key: &str) -> Option<String> {
-    let config = load_config()?;
-    let section = config.section(Some(section))?;
-    let value = section.get(key)?;
-
-    Some(value.to_string())
-}
-
-pub fn retrieve_folder_path() -> Option<String> {
-    retrieve_config_value("core", "path")
 }
 
 pub fn retrieve_db_config() -> DbConfig {
