+++ tests/cli/create/mod.rs
@@ -1,0 +1,3 @@
+mod event;
+mod migration;
+mod schema;
