+++ tests/helpers/mod.rs
@@ -1,9 +1,11 @@
 mod assert;
 mod common;
+mod config;
 mod io;
 mod surrealdb;
 
 pub use crate::helpers::surrealdb::*;
 pub use assert::*;
 pub use common::*;
+pub use config::*;
 pub use io::*;
